requirements.txt
backcall==0.2.0
click==8.0.1
decorator==5.0.9
Flask==2.0.1
greenlet==1.1.0
ipython==7.24.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
numpy==1.21.0
pandas==1.2.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.19
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==2.0.1
